added selection sort

+++ 2-selection_sort.c
@@ -10,7 +10,7 @@
 void selection_sort(int *array, size_t size)
 {
 	size_t i, j;
-	int min, temp;
+	size_t min, temp;
 
 	for (i = 0; i < size - 1; i++)
 	{
